+++ packages/extension-base/src/services/transaction-service/helpers/index.ts
@@ -84,6 +84,14 @@ const typeName = (type: SWTransactionBase['extrinsicType']) => {
       return 'Unlock votes';
     case ExtrinsicType.CHANGE_EARNING_VALIDATOR:
       return 'Change validator';
+    case ExtrinsicType.MULTISIG_APPROVE_TX:
+      return 'Approve multisig extrinsic';
+    case ExtrinsicType.MULTISIG_EXECUTE_TX:
+      return 'Execute multisig extrinsic';
+    case ExtrinsicType.MULTISIG_CANCEL_TX:
+      return 'Cancel multisig extrinsic';
+    case ExtrinsicType.MULTISIG_INIT_TX:
+      return 'Initiate multisig extrinsic';
     case ExtrinsicType.UNKNOWN:
     default:
       return 'unknown';