+++ packages/extension-koni-ui/src/Popup/Transaction/variants/Governance/ReferendumVote/ReferendumSplitVote.tsx
@@ -59,7 +59,7 @@ const Component = (props: ComponentProps): React.ReactElement<ComponentProps> =>
   const nayAmount = useWatchTransaction('nayAmount', form, formDefault);
 
   const { chainInfoMap } = useSelector((root) => root.chainStore);
-  const onPreCheck = usePreCheckAction(fromValue);
+  const onPreCheck = usePreCheckAction({ chain: chainValue, address: fromValue });
   const { onError, onSuccess } = useHandleSubmitTransaction();
 
   const { voteMap = new Map<string, ReferendumVoteDetail>(),
@@ -146,9 +146,7 @@ const Component = (props: ComponentProps): React.ReactElement<ComponentProps> =>
     };
 
     handleVote(voteRequest)
-      .then((tx) => {
-        onSuccess(tx);
-      })
+      .then(onSuccess)
       .catch(onError)
       .finally(() => setLoading(false));
   }, [chainValue, defaultData.referendumId, defaultData.track, onError, onSuccess]);