+++ packages/extension-koni-ui/src/components/Modal/Governance/GovVotingStatusModal/NestedVoteList.tsx
@@ -337,6 +337,10 @@ export const NestedVoteList = styled(forwardRef(Component))<Props>(({ theme: { t
 
       '&.-is-derived': {
         color: token.colorWarning
+      },
+
+      '&.-is-multisig': {
+        color: token['geekblue-9']
       }
     }
   };